+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## next
-
+* New showHeaderTabs prop that can be used to hide tabs in the header (e.g. in a single step wizard)
+* Footer's page indicators/navigation arrows are now hidden if wizard has only one step
+* Step name is no longer a required prop, because it is possible to hide header tabs altogether
 
 ## 1.1.1
 * Fixed obsolete class names: oc-mandatory and oc-mandatory-error.